Add registration/login
Added registration & login system. Added functions in login_scripts.js to send credentials to validate_form.jsp to validate/register user info on the fly and set the sesion attribute user, which we can use to determine if and who is logged in. Need to add a user home page or something to redirect user to after login